pyproject.toml
[project]
name = "notepadautomation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "botcity-framework-core>=1.3.0",
    "packaging>=25.0",
    "pynput>=1.8.1",
    "requests>=2.32.5",
]
